Add Confluent services to README
Questionable if we should add all the services to this list or only the
ones that are available to the host machine. I've opted to include
Kafka, Zookeeper and Schema Registry for now to announce them as they're
new. We could remove them in the future when they become well integrated
and expected.