Make can now target a single project
Before this change all operations moved across all projects.

1. all projects cloned, one at a time.
2. all projects pre-build hook, one at a time.
3. etc..

There wasn't a way to initialize a single project without initializing
all of them. That's not convenient when a new project is added to the
system.

This change allows the user to run `make init-newproject` to initialize
the single newproject.

In trade, the `network-create` target will run 1 time for each project.