Delimit `make list` output with linebreaks
This is a useful command to see the list of targets generated by the
meta templating. Before this change the targets were space delimited.
This outputs them with linebreaks is easier to grok and work with on the
command line.

For example, to see targets directly related to hydra:

    make list | grep hydra

Similarly, for network related tasks:

    make list | grep network