Extract config from Makefile and allow override
Renames configuration variable names to be more generic. Subprojects may
be official Reaction projects but you are free to add your own.

Moves configuration to the `config.mk` file which is included in the
Makefile. This provides an opportunity to include multiple files in
which configuration variables can be redefined.

As an alternative implementation, it would be possible to keep the
configuration in the main Makefile and include the secondary config.
This would be functionally equivalent. I think extracting both is better
as it provides a documented example for the configuration file.

The project's `config.mk` file can be copied to `config.local.mk` and
modified.