Improve gitignore coverage on customizations
Gitignore modifications that will ignore more appropriately when
customizing the project. For example, added subprojects and
configuration overrides will be automatically ignored. Root level files
and existing directories are not ignored.

Here's how it works:

Explicitly adds ignores for `config.local.mk` which can be added locally
to override the configuration that ships with the project in
`config.mk`.

Existing project directories are added as exclusions - meaning that they
and their contents will not be ignored. This ensures that changes to
these files will not be missed. Add new directories that should not be
ignored in a similar fashion.

Wildcard subdirectories will be ignored so that new subprojects don't
have to be explicitly added to the gitignore list.