Isolate automatic checkout to first clone
The platform configuration defines project, their git repos, and the git
ref (i.e. branch or tag) that should be used.

It could be surprising to perform a checkout of project branches in the
normal start operations. This change will automatically perform a
checkout after the initial clone.

The `checkout` target can be called manually after that.

--- Implementation ---

There is some redundancy to perform the checkout in the clone and the
standalone checkout target. I'd hoped to abstract the folder target but
when adding dependencies, the dependencies still ran.

There's probably a better way to do this with Makefile-fu, but this
seems to work OK!