feature: Update checkouts
Adds the ability to update the local project state to the branches
configured in the config file.

If you have done any local development or checked out branches, then
this will:

1. Ensure there is no uncommitted work in any sub-projects, else the
operation will fail with a message to let you know which project needs
attention.
2. Checkout the project branch specified in the config file.
3. Pull the checked out branch from origin. The name `origin` is
hardcoded at this time. The pull is configured with `--ff-only` and will
fail if a fast-forward is not possible.

No breaking changes. No existing behavior is changed.

Signed-off-by: Ticean Bennett <ticean@reactioncommerce.com>